feat: whisper 스펙 변경으로 인해 수정

+++ src/main/java/com/jolupbisang/demo/application/meeting/event/AgendaReceivedEvent.java
@@ -8,9 +8,9 @@ import java.util.List;
 @Getter
 public class AgendaReceivedEvent extends ApplicationEvent {
     private final long meetingId;
-    private final List<Integer> agenda;
+    private final List<Long> agenda;
 
-    public AgendaReceivedEvent(Object source, long meetingId, List<Integer> agenda) {
+    public AgendaReceivedEvent(Object source, long meetingId, List<Long> agenda) {
         super(source);
         this.meetingId = meetingId;
         this.agenda = agenda;
